Add in `git-fork` script
This is just a script to use the GitHub API to fork a repository and
then clone it down and setup the remotes to define both the forked
repository and the original (using origin and upstream respectively).

Written in pure shell script because I hate myself.  Relies on `curl`
for composing the HTTP requests and `jq` for parsing the response
bodies.